perf: defer Octokit/auth from home route and harden CLS on marketing surfaces
Splits the home bundle so `/` no longer preloads Octokit, and removes the
highest-confidence CLS risks on shared marketing surfaces.

- Root providers reduced to theme-only; GitHubDataProvider is now scoped to
  the dashboard route, which is lazy-loaded via React.lazy + Suspense.
- Header split into MarketingHeader (no useGitHubData, uses new
  useStoredPatPresence) and DashboardHeader; dashboard header lives inside
  the dashboard shell.
- PAT form extracted to a lazy GetStartedForm loaded via IntersectionObserver
  and a preload event dispatched by the hero CTA; section renders a fixed-
  height fallback to prevent shift during hydration.
- Form persists PAT through secureStorage and a sessionStorage ephemeral
  flag; provider honors the flag on mount so "Remember my token = off"
  still clears the stored token.
- product-showcase images now sit in fixed aspect-ratio wrappers with
  intrinsic width/height + loading=lazy + decoding=async; footer
  ProductHunt badge wrapped in a fixed 150x32 box.
- vite.config manualChunks no longer splits @octokit (shared babel-runtime
  was leaking it into the entry); Rollup auto-extracts a shared chunk that
  loads only with the dashboard or form.
- Dev-only CLS logger at ?debug-layout-shift=1 using PerformanceObserver.
- Tests: updated header/get-started-section specs, added product-showcase
  test, added e2e/home-bundle.spec.ts asserting no Octokit on initial /.
