+++ src/app.tsx
@@ -1,14 +1,20 @@
 import { MotionConfig } from "framer-motion";
-import { useEffect } from "react";
+import { lazy, Suspense, useEffect } from "react";
 import { Navigate, Route, Routes, useLocation } from "react-router-dom";
 
 import { ErrorBoundary } from "@/components/error-boundary";
 import { FathomAnalytics } from "@/components/fathom-analytics";
 import Footer from "@/components/footer";
-import Header from "@/components/header";
+import { MarketingHeader } from "@/components/marketing-header";
 import { Providers } from "@/providers/providers";
-import { Dashboard } from "@/routes/dashboard";
 import { Home } from "@/routes/home";
+
+// Dashboard subtree is lazy so `/` does not ship Octokit / auth code.
+const DashboardShell = lazy(() =>
+  import("@/routes/dashboard-shell").then((m) => ({
+    default: m.DashboardShell,
+  })),
+);
 
 /** Reset scroll position on route change */
 function ScrollToTop() {
@@ -19,6 +25,16 @@ function ScrollToTop() {
   return null;
 }
 
+function MarketingLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <div className="min-h-full">
+      <MarketingHeader />
+      <main>{children}</main>
+      <Footer />
+    </div>
+  );
+}
+
 export function App() {
   return (
     <Providers themeProps={{ attribute: "class", defaultTheme: "light" }}>
@@ -26,24 +42,25 @@ export function App() {
         <ErrorBoundary>
           <FathomAnalytics />
           <ScrollToTop />
-          <div className="min-h-full">
-            <Header />
-            <main>
-              <Routes>
-                <Route element={<Home />} path="/" />
-                <Route
-                  element={
-                    <div className="max-w-7xl mx-auto px-4 sm:px-6">
-                      <Dashboard />
-                    </div>
-                  }
-                  path="/dashboard"
-                />
-                <Route element={<Navigate replace to="/" />} path="*" />
-              </Routes>
-            </main>
-            <Footer />
-          </div>
+          <Routes>
+            <Route
+              element={
+                <MarketingLayout>
+                  <Home />
+                </MarketingLayout>
+              }
+              path="/"
+            />
+            <Route
+              element={
+                <Suspense fallback={null}>
+                  <DashboardShell />
+                </Suspense>
+              }
+              path="/dashboard"
+            />
+            <Route element={<Navigate replace to="/" />} path="*" />
+          </Routes>
         </ErrorBoundary>
       </MotionConfig>
     </Providers>
