+++ src/components/footer.tsx
@@ -38,13 +38,16 @@ export default function Footer() {
               </a>
               <a
                 aria-label="Repo Remover on Product Hunt"
+                className="block"
                 href="https://www.producthunt.com/products/repo-remover/launches/repo-remover?embed=true&utm_source=badge-featured&utm_medium=badge&utm_campaign=badge-repo-remover"
                 rel="noopener noreferrer"
+                style={{ height: 32, width: 150 }}
                 target="_blank"
               >
                 <img
                   alt="Repo Remover - Archive or delete multiple GitHub repos with a single click. | Product Hunt"
-                  className="h-8 w-auto"
+                  className="block h-8 w-[150px]"
+                  decoding="async"
                   height={32}
                   loading="lazy"
                   src="https://api.producthunt.com/widgets/embed-image/v1/featured.svg?post_id=159117&theme=light&t=1775226012179"
