+++ src/components/landing/get-started-section.tsx
@@ -1,33 +1,8 @@
 import { motion, useReducedMotion } from "framer-motion";
-import {
-  Check,
-  ExternalLink,
-  Eye,
-  EyeOff,
-  Info,
-  Key,
-  LoaderCircle,
-  Search,
-  SquareCheckBig,
-} from "lucide-react";
-import { useState } from "react";
-import { useNavigate } from "react-router-dom";
+import { Check, ExternalLink, Key, Search, SquareCheckBig } from "lucide-react";
+import { lazy, Suspense, useEffect, useRef, useState } from "react";
 
-import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
-import { Button } from "@/components/ui/button";
 import { Card, CardContent } from "@/components/ui/card";
-import { Checkbox } from "@/components/ui/checkbox";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-
-import { isValidGitHubToken } from "@/github/client";
-import { useGitHubData } from "@/hooks/use-github-data";
-import { useTokenValidation } from "@/hooks/use-token-validation";
 import { analytics } from "@/utils/analytics";
 import {
   fadeUp,
@@ -36,6 +11,16 @@ import {
   staggerContainer,
   staggerContainerWide,
 } from "@/utils/motion";
+
+/**
+ * Custom event signalling that the user is heading toward the Get Started
+ * section (e.g. the hero CTA). The section listens and hydrates the
+ * interactive form ahead of the intersection observer firing so the input is
+ * ready when the user arrives.
+ */
+export const PRELOAD_GET_STARTED_FORM_EVENT = "preload-get-started-form";
+
+const GetStartedForm = lazy(() => import("./get-started-form"));
 
 const steps = [
   {
@@ -66,148 +51,78 @@ const steps = [
 ];
 
 /**
- * Inline PAT form — matches Figma design (key icon + password input + full-width button).
- * Handles validation + auth without HeroUI components.
+ * Placeholder sized to match the interactive form so the card does not resize
+ * when `GetStartedForm` mounts. Heights chosen to mirror the stacked layout:
+ * input + error/success line + checkbox row + submit button.
  */
-function InlinePATForm() {
-  const [token, setToken] = useState("");
-  const [remember, setRemember] = useState(true);
-  const [showToken, setShowToken] = useState(false);
-  const { setPat } = useGitHubData();
-  const navigate = useNavigate();
-  const { error, isValid, isValidating, scopeWarnings } =
-    useTokenValidation(token);
-
-  function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    if (!isValid || isValidating) return;
-
-    setPat(token, remember);
-    void navigate("/dashboard");
-  }
-
-  const canSubmit = isValid && !isValidating;
-  const formatError =
-    token.length > 5 && !isValidGitHubToken(token)
-      ? "Invalid token format"
-      : null;
-  const displayError = error ?? formatError;
-
+function FormFallback() {
   return (
-    <form className="space-y-4" onSubmit={handleSubmit}>
-      <div className="relative">
-        <Key className="absolute left-3 top-1/2 -translate-y-1/2 h-4 w-4 text-default-400" />
-        <Input
-          autoComplete="off"
-          className={`h-auto pl-10 pr-12 py-3 bg-default-100 text-sm font-mono placeholder:text-default-400 ${
-            displayError
-              ? "border-danger"
-              : isValid
-                ? "border-emerald-600"
-                : "border-divider"
-          }`}
-          data-testid="github-token-input"
-          onChange={(e) => setToken(e.target.value)}
-          placeholder="ghp_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
-          type={showToken ? "text" : "password"}
-          data-1p-ignore
-          data-lpignore="true"
-          value={token}
-        />
-        <div className="absolute right-3 top-1/2 -translate-y-1/2 flex items-center gap-1">
-          {token && (
-            <Button
-              aria-label={showToken ? "Hide token" : "Show token"}
-              className="text-default-400 hover:text-default-600 transition-colors p-0.5"
-              onClick={() => setShowToken((prev) => !prev)}
-              variant="ghost"
-              size="icon"
-            >
-              {showToken ? (
-                <EyeOff className="h-4 w-4" />
-              ) : (
-                <Eye className="h-4 w-4" />
-              )}
-            </Button>
-          )}
-          {isValidating && (
-            <LoaderCircle className="h-4 w-4 text-default-400 animate-spin" />
-          )}
-        </div>
-      </div>
-
-      {displayError && <p className="text-xs text-danger">{displayError}</p>}
-
-      {isValid && (
-        <p className="text-xs text-emerald-700 dark:text-emerald-400">
-          Token validated successfully
-        </p>
-      )}
-
-      {scopeWarnings.length > 0 && (
-        <Alert variant="warning" data-testid="scope-warnings">
-          <AlertTitle>Missing recommended scopes:</AlertTitle>
-          <AlertDescription>
-            <ul className="mt-1 list-inside list-disc space-y-0.5 text-xs">
-              {scopeWarnings.map((warning) => (
-                <li key={warning}>{warning}</li>
-              ))}
-            </ul>
-            <a
-              className="mt-2 inline-block text-xs font-medium underline hover:no-underline"
-              href="https://github.com/settings/tokens"
-              rel="noopener noreferrer"
-              target="_blank"
-            >
-              Update token scopes on GitHub &rarr;
-            </a>
-          </AlertDescription>
-        </Alert>
-      )}
-
+    <div className="space-y-4" data-testid="get-started-form-fallback">
+      <div className="h-[46px] w-full rounded-md bg-default-100 border border-divider" />
+      <div className="h-4" />
       <div className="flex items-center gap-2">
-        <Checkbox
-          checked={remember}
-          data-testid="github-token-remember"
-          id="remember-token"
-          onCheckedChange={(checked) => setRemember(checked)}
-        />
-        <Label className="text-default-500" htmlFor="remember-token">
-          Remember my token
-        </Label>
-        <Popover>
-          <PopoverTrigger
-            aria-label="Token storage info"
-            className="inline-flex h-8 w-8 items-center justify-center rounded-md text-default-400 hover:text-default-600 hover:bg-content2 transition-colors cursor-pointer"
-          >
-            <Info className="h-3.5 w-3.5" />
-          </PopoverTrigger>
-          <PopoverContent side="top" className="w-64 text-xs text-default-500">
-            Your token is encrypted locally before being saved to your browser.
-            It is never sent to a server.
-          </PopoverContent>
-        </Popover>
+        <div className="h-4 w-4 rounded-sm bg-default-100 border border-divider" />
+        <div className="h-4 w-40 rounded bg-default-100" />
       </div>
-
-      <Button
-        className={`w-full h-12 text-base ${
-          canSubmit
-            ? "bg-[var(--brand-blue)] text-white hover:opacity-90 shadow-sm"
-            : "bg-default-200 text-default-500 cursor-not-allowed"
-        }`}
-        data-testid="github-token-submit"
-        disabled={!canSubmit}
-        size="lg"
-        type="submit"
-      >
-        {isValidating ? "Validating..." : "Load My Repositories"}
-      </Button>
-    </form>
+      <div className="h-12 w-full rounded-md bg-default-200" />
+    </div>
   );
+}
+
+function useDeferredFormLoad(): [
+  boolean,
+  React.RefObject<HTMLDivElement | null>,
+] {
+  const containerRef = useRef<HTMLDivElement | null>(null);
+  const [shouldLoad, setShouldLoad] = useState(false);
+
+  useEffect(() => {
+    if (shouldLoad) return;
+
+    const load = () => setShouldLoad(true);
+
+    // Preload signal from the hero CTA — fires immediately so the form is
+    // hydrating while the smooth scroll is still in flight.
+    window.addEventListener(PRELOAD_GET_STARTED_FORM_EVENT, load, {
+      once: true,
+    });
+
+    // Intersection observer as the fallback trigger for users who scroll
+    // organically past the section.
+    let observer: IntersectionObserver | undefined;
+    const node = containerRef.current;
+    if (node && typeof IntersectionObserver !== "undefined") {
+      observer = new IntersectionObserver(
+        (entries) => {
+          for (const entry of entries) {
+            if (entry.isIntersecting) {
+              load();
+              observer?.disconnect();
+              break;
+            }
+          }
+        },
+        { rootMargin: "400px 0px" },
+      );
+      observer.observe(node);
+    } else {
+      // Environment without IntersectionObserver — load immediately so the
+      // form is still reachable (e.g. SSR/older browsers, tests).
+      load();
+    }
+
+    return () => {
+      window.removeEventListener(PRELOAD_GET_STARTED_FORM_EVENT, load);
+      observer?.disconnect();
+    };
+  }, [shouldLoad]);
+
+  return [shouldLoad, containerRef];
 }
 
 export function GetStartedSection() {
   const reduced = useReducedMotion();
+  const [shouldLoadForm, formContainerRef] = useDeferredFormLoad();
 
   return (
     <section className="w-full px-6 sm:px-8 py-16 sm:py-20" id="get-started">
@@ -277,7 +192,7 @@ export function GetStartedSection() {
           </div>
         </motion.div>
 
-        {/* PAT form card */}
+        {/* PAT form card — deferred so the home bundle omits Octokit */}
         <motion.div
           className="mt-16 max-w-xl mx-auto"
           {...scrollRevealProps(scaleIn, reduced)}
@@ -291,7 +206,15 @@ export function GetStartedSection() {
                 Paste your PAT below to load your repositories.
               </p>
 
-              <InlinePATForm />
+              <div ref={formContainerRef}>
+                {shouldLoadForm ? (
+                  <Suspense fallback={<FormFallback />}>
+                    <GetStartedForm />
+                  </Suspense>
+                ) : (
+                  <FormFallback />
+                )}
+              </div>
 
               <div className="flex justify-center mt-5">
                 <div className="flex flex-col sm:flex-row items-start sm:items-center gap-1.5 sm:gap-6 text-xs text-default-400">
