+++ src/providers/providers.tsx
@@ -4,17 +4,16 @@ import {
 } from "next-themes";
 import { type ReactNode } from "react";
 
-import { GitHubDataProvider } from "@/providers/github-data-provider";
-
 export interface ProvidersProps {
   children: ReactNode;
   themeProps?: ThemeProviderProps;
 }
 
+/**
+ * App-root providers. Intentionally theme-only so the home route can ship
+ * without Octokit / GitHub auth code. `GitHubDataProvider` is mounted inside
+ * the dashboard route subtree.
+ */
 export function Providers({ children, themeProps }: ProvidersProps) {
-  return (
-    <NextThemesProvider {...themeProps}>
-      <GitHubDataProvider>{children}</GitHubDataProvider>
-    </NextThemesProvider>
-  );
+  return <NextThemesProvider {...themeProps}>{children}</NextThemesProvider>;
 }
